#41 review fix
